refactor(types): единый загрузчик JSON-паков платформенных типов
Парсинг встроенных JSON-паков платформенных типов вынесен из провайдеров
в общий BuiltinTypesJsonLoader. И BSL-, и OneScript-провайдеры грузят одну
и ту же структуру ресурса — теперь через один загрузчик, а не через
static-метод BSL-провайдера, в который лез OScript-провайдер.

BuiltinPlatformTypesProvider и BuiltinOScriptPlatformTypesProvider оставляют
только провайдерную логику (путь ресурса, кэш, языковой скоуп, gating по
bsl-context). Логика разбора (load/readMembers/readSignatures/readMetadata/
readTypeSet/...) перенесена без изменений поведения.

Ссылки в тестах обновлены на BuiltinTypesJsonLoader.load.
